ha-vmi-addon-export/requirements.txt
pyserial>=3.5
paho-mqtt>=1.7.0
enocean>=0.60.1
